revert

+++ src/bp/core/services/ghost/service.ts
@@ -1,9 +1,7 @@
-import { ListenHandle, Logger, UpsertOptions } from 'botpress/sdk'
+import { ListenHandle, Logger } from 'botpress/sdk'
 import { ObjectCache } from 'common/object-cache'
 import { isValidBotId } from 'common/validation'
-import { BotConfig } from 'core/config/bot.config'
-import { asBytes, filterByGlobs, forceForwardSlashes } from 'core/misc/utils'
-import { diffLines } from 'diff'
+import { asBytes, forceForwardSlashes } from 'core/misc/utils'
 import { EventEmitter2 } from 'eventemitter2'
 import fse from 'fs-extra'
 import { inject, injectable, tagged } from 'inversify'
@@ -18,32 +16,11 @@ import { VError } from 'verror'
 import { createArchive } from '../../misc/archive'
 import { TYPES } from '../../types'
 
-import { FileRevision, PendingRevisions, ReplaceContent, ServerWidePendingRevisions, StorageDriver } from '.'
+import { PendingRevisions, ReplaceContent, ServerWidePendingRevisions, StorageDriver } from '.'
 import DBStorageDriver from './db-driver'
 import DiskStorageDriver from './disk-driver'
 
-export type BpfsScopedChange = {
-  // An undefined bot ID = global
-  botId: string | undefined
-  // The list of local files which will overwrite their remote counterpart
-  localFiles: string[]
-  // List of added/deleted files based on local and remote files, and differences between files from revisions
-  changes: FileChange[]
-}
-
-export interface FileChange {
-  path: string
-  action: FileChangeAction
-  add?: number
-  del?: number
-}
-
-export type FileChangeAction = 'add' | 'edit' | 'del'
-
 const MAX_GHOST_FILE_SIZE = asBytes('100mb')
-const bpfsIgnoredFiles = ['models/**', 'data/bots/*/models/**', '**/*.js.map']
-const GLOBAL_GHOST_KEY = '__global__'
-const BOTS_GHOST_KEY = '__bots__'
 
 @injectable()
 export class GhostService {
@@ -57,26 +34,14 @@ export class GhostService {
     @inject(TYPES.Logger)
     @tagged('name', 'GhostService')
     private logger: Logger
-  ) {
-    this.cache.events.on && this.cache.events.on('syncDbFilesToDisk', this._onSyncReceived)
-  }
+  ) {}
 
   initialize(enabled: boolean) {
     this.enabled = enabled
-    this._scopedGhosts.clear()
-  }
-
-  // Not caching this scope since it's rarely used
-  root(): ScopedGhostService {
-    return new ScopedGhostService(`./data`, this.diskDriver, this.dbDriver, this.enabled, this.cache, this.logger)
   }
 
   global(): ScopedGhostService {
-    if (this._scopedGhosts.has(GLOBAL_GHOST_KEY)) {
-      return this._scopedGhosts.get(GLOBAL_GHOST_KEY)!
-    }
-
-    const scopedGhost = new ScopedGhostService(
+    return new ScopedGhostService(
       `./data/global`,
       this.diskDriver,
       this.dbDriver,
@@ -84,140 +49,10 @@ export class GhostService {
       this.cache,
       this.logger
     )
-
-    this._scopedGhosts.set(GLOBAL_GHOST_KEY, scopedGhost)
-    return scopedGhost
-  }
-
-  custom(baseDir: string) {
-    return new ScopedGhostService(baseDir, this.diskDriver, this.dbDriver, false, this.cache, this.logger)
-  }
-
-  // TODO: refactor this
-  async forceUpdate(tmpFolder: string) {
-    const invalidateFile = async (fileName: string) => {
-      await this.cache.invalidate(`object::${fileName}`)
-      await this.cache.invalidate(`buffer::${fileName}`)
-    }
-
-    const dbRevs = await this.dbDriver.listRevisions('data/')
-    await Promise.each(dbRevs, rev => this.dbDriver.deleteRevision(rev.path, rev.revision))
-
-    const allChanges = await this.listFileChanges(tmpFolder)
-    for (const { changes, localFiles } of allChanges) {
-      await Promise.map(changes.filter(x => x.action === 'del'), async file => {
-        await this.dbDriver.deleteFile(file.path)
-        await invalidateFile(file.path)
-      })
-
-      // Upload all local files for that scope
-      if (localFiles.length) {
-        await Promise.map(localFiles, async filePath => {
-          const content = await this.diskDriver.readFile(path.join(tmpFolder, filePath))
-          await this.dbDriver.upsertFile(filePath, content, false)
-          await invalidateFile(filePath)
-        })
-      }
-    }
-
-    return allChanges.filter(x => x.localFiles.length && x.botId).map(x => x.botId)
-  }
-
-  // TODO: refactor this
-  async listFileChanges(tmpFolder: string): Promise<BpfsScopedChange[]> {
-    const tmpDiskGlobal = this.custom(path.resolve(tmpFolder, 'data/global'))
-    const tmpDiskBot = (botId?: string) => this.custom(path.resolve(tmpFolder, 'data/bots', botId || ''))
-
-    // We need local and remote bot ids to correctly display changes
-    const remoteBotIds = (await this.bots().directoryListing('/', 'bot.config.json')).map(path.dirname)
-    const localBotIds = (await tmpDiskBot().directoryListing('/', 'bot.config.json')).map(path.dirname)
-    const botsIds = _.uniq([...remoteBotIds, ...localBotIds])
-
-    const uniqueFile = file => `${file.path} | ${file.revision}`
-
-    const getFileDiff = async (file: string): Promise<FileChange> => {
-      try {
-        const localFile = (await this.diskDriver.readFile(path.join(tmpFolder, file))).toString()
-        const dbFile = (await this.dbDriver.readFile(file)).toString()
-
-        const diff = diffLines(dbFile, localFile)
-
-        return {
-          path: file,
-          action: 'edit' as FileChangeAction,
-          add: _.sumBy(diff.filter(d => d.added), 'count'),
-          del: _.sumBy(diff.filter(d => d.removed), 'count')
-        }
-      } catch (err) {
-        // Todo better handling
-        this.logger.attachError(err).error(`Error while checking diff for "${file}"`)
-        return { path: file, action: 'edit' as FileChangeAction }
-      }
-    }
-
-    // Adds the correct prefix to files so they are displayed correctly when reviewing changes
-    const getDirectoryFullPaths = async (botId: string | undefined, ghost: ScopedGhostService) => {
-      const getPath = (file: string) => (botId ? path.join('data/bots', botId, file) : path.join('data/global', file))
-
-      return ghost
-        .directoryListing('/', '*.*', [...bpfsIgnoredFiles, '**/revisions.json'])
-        .then()
-        .map(f => forceForwardSlashes(getPath(f)))
-    }
-
-    const filterRevisions = (revisions: FileRevision[]) => filterByGlobs(revisions, r => r.path, bpfsIgnoredFiles)
-
-    const getFileChanges = async (
-      botId: string | undefined,
-      localGhost: ScopedGhostService,
-      remoteGhost: ScopedGhostService
-    ) => {
-      const localRevs = filterRevisions(await localGhost.listDiskRevisions())
-      const remoteRevs = filterRevisions(await remoteGhost.listDbRevisions())
-      const syncedRevs = _.intersectionBy(localRevs, remoteRevs, uniqueFile)
-      const unsyncedFiles = _.uniq(_.differenceBy(remoteRevs, syncedRevs, uniqueFile).map(x => x.path))
-
-      const localFiles: string[] = await getDirectoryFullPaths(botId, localGhost)
-      const remoteFiles: string[] = await getDirectoryFullPaths(botId, remoteGhost)
-
-      const deleted = _.difference(remoteFiles, localFiles).map(x => ({ path: x, action: 'del' as FileChangeAction }))
-      const added = _.difference(localFiles, remoteFiles).map(x => ({ path: x, action: 'add' as FileChangeAction }))
-
-      const filterDeleted = file => !_.map([...deleted, ...added], 'path').includes(file)
-      const edited = (await Promise.map(unsyncedFiles.filter(filterDeleted), getFileDiff)).filter(
-        x => x.add !== 0 || x.del !== 0
-      )
-
-      return {
-        botId,
-        changes: [...added, ...deleted, ...edited],
-        localFiles
-      }
-    }
-
-    const botsFileChanges = await Promise.map(botsIds, botId =>
-      getFileChanges(botId, tmpDiskBot(botId), this.forBot(botId))
-    )
-
-    return [...botsFileChanges, await getFileChanges(undefined, tmpDiskGlobal, this.global())]
   }
 
   bots(): ScopedGhostService {
-    if (this._scopedGhosts.has(BOTS_GHOST_KEY)) {
-      return this._scopedGhosts.get(BOTS_GHOST_KEY)!
-    }
-
-    const scopedGhost = new ScopedGhostService(
-      `./data/bots`,
-      this.diskDriver,
-      this.dbDriver,
-      this.enabled,
-      this.cache,
-      this.logger
-    )
-
-    this._scopedGhosts.set(BOTS_GHOST_KEY, scopedGhost)
-    return scopedGhost
+    return new ScopedGhostService(`./data/bots`, this.diskDriver, this.dbDriver, this.enabled, this.cache, this.logger)
   }
 
   forBot(botId: string): ScopedGhostService {
@@ -235,8 +70,7 @@ export class GhostService {
       this.dbDriver,
       this.enabled,
       this.cache,
-      this.logger,
-      botId
+      this.logger
     )
 
     const listenForUnmount = args => {
@@ -252,27 +86,25 @@ export class GhostService {
     return scopedGhost
   }
 
-  public async exportArchive(): Promise<Buffer> {
+  public async exportArchive(botIds: string[]): Promise<Buffer> {
     const tmpDir = tmp.dirSync({ unsafeCleanup: true })
-
-    const getFullPath = folder => path.join(tmpDir.name, folder)
+    const files: string[] = []
 
     try {
-      const botIds = (await this.bots().directoryListing('/', 'bot.config.json')).map(path.dirname)
-      const botFiles = await Promise.mapSeries(botIds, async botId =>
-        (await this.forBot(botId).exportToDirectory(getFullPath(`bots/${botId}`), bpfsIgnoredFiles)).map(f =>
-          path.join(`bots/${botId}`, f)
-        )
-      )
+      await mkdirp.sync(path.join(tmpDir.name, 'global'))
+      const outDir = path.join(tmpDir.name, 'global')
+      const outFiles = (await this.global().exportToDirectory(outDir)).map(f => path.join('global', f))
+      files.push(...outFiles)
 
-      const allFiles = [
-        ..._.flatten(botFiles),
-        ...(await this.global().exportToDirectory(getFullPath('global'), bpfsIgnoredFiles)).map(f =>
-          path.join('global', f)
-        )
-      ]
+      await Promise.mapSeries(botIds, async bid => {
+        const p = path.join(tmpDir.name, `bots/${bid}`)
+        await mkdirp.sync(p)
+        const outFiles = (await this.forBot(bid).exportToDirectory(p)).map(f => path.join(`bots/${bid}`, f))
+        files.push(...outFiles)
+      })
 
-      const archive = await createArchive(getFullPath('archive.tgz'), tmpDir.name, allFiles)
+      const filename = path.join(tmpDir.name, 'archive.tgz')
+      const archive = await createArchive(filename, tmpDir.name, files)
       return await fse.readFile(archive)
     } finally {
       tmpDir.removeCallback()
@@ -289,19 +121,6 @@ export class GhostService {
     return {
       global,
       bots
-    }
-  }
-
-  private _onSyncReceived = async (message: string) => {
-    try {
-      const { rootFolder, botId } = JSON.parse(message)
-      if (botId) {
-        await this.forBot(botId).syncDatabaseFilesToDisk(rootFolder)
-      } else {
-        await this.global().syncDatabaseFilesToDisk(rootFolder)
-      }
-    } catch (err) {
-      this.logger.attachError(err).error('Could not sync files locally.')
     }
   }
 }
@@ -322,8 +141,7 @@ export class ScopedGhostService {
     private dbDriver: DBStorageDriver,
     private useDbDriver: boolean,
     private cache: ObjectCache,
-    private logger: Logger,
-    private botId?: string
+    private logger: Logger
   ) {
     if (![-1, this.baseDir.length - 1].includes(this.baseDir.indexOf('*'))) {
       throw new Error(`Base directory can only contain '*' at the end of the path`)
@@ -331,23 +149,6 @@ export class ScopedGhostService {
 
     this.isDirectoryGlob = this.baseDir.endsWith('*')
     this.primaryDriver = useDbDriver ? dbDriver : diskDriver
-  }
-
-  /**
-   * TODO: Refactor this on v12.1.4
-   * This is a temporary workaround to lock bots marked as "locked" until modules are correctly updated.
-   */
-  private async _assertBotUnlocked(directory: string, file?: string) {
-    if (!this.botId || directory.startsWith('./models')) {
-      return
-    }
-
-    if (await this.fileExists('/', 'bot.config.json')) {
-      const config = await this.readFileAsObject<BotConfig>('/', 'bot.config.json')
-      if (config.locked) {
-        throw new Error(`Bot locked`)
-      }
-    }
   }
 
   private _normalizeFolderName(rootFolder: string) {
@@ -377,20 +178,7 @@ export class ScopedGhostService {
     }
   }
 
-  async upsertFile(
-    rootFolder: string,
-    file: string,
-    content: string | Buffer,
-    options: UpsertOptions = {
-      recordRevision: true,
-      syncDbToDisk: false,
-      ignoreLock: false
-    }
-  ): Promise<void> {
-    if (!options.ignoreLock) {
-      await this._assertBotUnlocked(rootFolder, file)
-    }
-
+  async upsertFile(rootFolder: string, file: string, content: string | Buffer, recordRevision = true): Promise<void> {
     if (this.isDirectoryGlob) {
       throw new Error(`Ghost can't read or write under this scope`)
     }
@@ -401,25 +189,17 @@ export class ScopedGhostService {
       throw new Error(`The size of the file ${fileName} is over the 100mb limit`)
     }
 
-    await this.primaryDriver.upsertFile(fileName, content, !!options.recordRevision)
+    await this.primaryDriver.upsertFile(fileName, content, recordRevision)
     this.events.emit('changed', fileName)
     await this._invalidateFile(fileName)
-
-    if (options.syncDbToDisk) {
-      await this.cache.sync(JSON.stringify({ rootFolder, botId: this.botId }))
-    }
   }
 
-  async upsertFiles(rootFolder: string, content: FileContent[], options?: UpsertOptions): Promise<void> {
-    if (options && !options.ignoreLock) {
-      await this._assertBotUnlocked(rootFolder)
-    }
-
+  async upsertFiles(rootFolder: string, content: FileContent[]): Promise<void> {
     await Promise.all(content.map(c => this.upsertFile(rootFolder, c.name, c.content)))
   }
 
   /**
-   * Sync the local filesystem to the database.
+   * All tracked files will be synced.
    * All files are tracked by default, unless `.ghostignore` is used to exclude them.
    */
   async sync() {
@@ -428,25 +208,32 @@ export class ScopedGhostService {
       return
     }
 
-    const localFiles = await this.diskDriver.directoryListing(this.baseDir, { includeDotFiles: true })
+    // Get files from disk that should be ghosted
+    const trackedFiles = await this.diskDriver.directoryListing(this.baseDir, { includeDotFiles: true })
+
     const diskRevs = await this.diskDriver.listRevisions(this.baseDir)
     const dbRevs = await this.dbDriver.listRevisions(this.baseDir)
     const syncedRevs = _.intersectionBy(diskRevs, dbRevs, x => `${x.path} | ${x.revision}`)
 
     await Promise.each(syncedRevs, rev => this.dbDriver.deleteRevision(rev.path, rev.revision))
-    await this._updateProduction(localFiles)
-  }
 
-  private async _updateProduction(localFiles: string[]) {
-    // Delete the prod files that has been deleted from disk
-    const prodFiles = await this.dbDriver.directoryListing(this._normalizeFolderName('./'))
-    const filesToDelete = _.difference(prodFiles, localFiles)
+    if (!(await this.isFullySynced())) {
+      const scUrl = `/admin/settings/version`
+      this.logger.warn(
+        `You have changes on your production environment that aren't synced on your local file system. Visit '${scUrl}' to save changes back to your Source Control.`
+      )
+      return
+    }
+
+    // Delete the ghosted files that has been deleted from disk
+    const ghostedFiles = await this.dbDriver.directoryListing(this._normalizeFolderName('./'))
+    const filesToDelete = _.difference(ghostedFiles, trackedFiles)
     await Promise.map(filesToDelete, filePath =>
       this.dbDriver.deleteFile(this._normalizeFileName('./', filePath), false)
     )
 
-    // Overwrite all of the prod files with the local files
-    await Promise.each(localFiles, async file => {
+    // Overwrite all of the ghosted files with the tracked files
+    await Promise.each(trackedFiles, async file => {
       const filePath = this._normalizeFileName('./', file)
       const content = await this.diskDriver.readFile(filePath)
       await this.dbDriver.upsertFile(filePath, content, false)
@@ -463,9 +250,11 @@ export class ScopedGhostService {
       await fse.writeFile(outPath, content)
     }
 
-    const dbRevs = await this.dbDriver.listRevisions(this.baseDir)
+    const oldRevisions = await this.diskDriver.listRevisions(this.baseDir)
+    const newRevisions = await this.dbDriver.listRevisions(this.baseDir)
+    const mergedRevisions = _.unionBy(oldRevisions, newRevisions, x => x.path + ' ' + x.revision)
 
-    await fse.writeFile(path.join(directory, 'revisions.json'), JSON.stringify(dbRevs, undefined, 2))
+    await fse.writeFile(path.join(directory, 'revisions.json'), JSON.stringify(mergedRevisions, undefined, 2))
     if (!allFiles.includes('revisions.json')) {
       allFiles.push('revisions.json')
     }
@@ -483,7 +272,7 @@ export class ScopedGhostService {
       } as FileContent
     })
 
-    await this.upsertFiles('/', files, { ignoreLock: true })
+    await this.upsertFiles('/', files)
   }
 
   public async exportToArchiveBuffer(exludes?: string | string[], replaceContent?: ReplaceContent): Promise<Buffer> {
@@ -559,7 +348,6 @@ export class ScopedGhostService {
   }
 
   async deleteFile(rootFolder: string, file: string): Promise<void> {
-    await this._assertBotUnlocked(rootFolder, file)
     if (this.isDirectoryGlob) {
       throw new Error(`Ghost can't read or write under this scope`)
     }
@@ -571,28 +359,13 @@ export class ScopedGhostService {
   }
 
   async renameFile(rootFolder: string, fromName: string, toName: string): Promise<void> {
-    await this._assertBotUnlocked(rootFolder, fromName)
     const fromPath = this._normalizeFileName(rootFolder, fromName)
     const toPath = this._normalizeFileName(rootFolder, toName)
 
     await this.primaryDriver.moveFile(fromPath, toPath)
   }
 
-  async syncDatabaseFilesToDisk(rootFolder: string): Promise<void> {
-    if (!this.useDbDriver) {
-      return
-    }
-
-    const remoteFiles = await this.dbDriver.directoryListing(this._normalizeFolderName(rootFolder))
-    const filePath = filename => this._normalizeFileName(rootFolder, filename)
-
-    await Promise.mapSeries(remoteFiles, async file =>
-      this.diskDriver.upsertFile(filePath(file), await this.dbDriver.readFile(filePath(file)))
-    )
-  }
-
   async deleteFolder(folder: string): Promise<void> {
-    await this._assertBotUnlocked(folder)
     if (this.isDirectoryGlob) {
       throw new Error(`Ghost can't read or write under this scope`)
     }
@@ -644,14 +417,6 @@ export class ScopedGhostService {
     }
 
     return result
-  }
-
-  async listDbRevisions(): Promise<FileRevision[]> {
-    return this.dbDriver.listRevisions(this.baseDir)
-  }
-
-  async listDiskRevisions(): Promise<FileRevision[]> {
-    return this.diskDriver.listRevisions(this.baseDir)
   }
 
   onFileChanged(callback: (filePath: string) => void): ListenHandle {
